upnpdevice: refactor interface
There is no need to have a UPnPDevice as a class. The devices need to be
passed to ContextManager later on, so keeping a reference to a
RootDevice within a class object makes no sense. Since we're not doing
any operations on a device anyway, convert the class to a namespace with
static methods only.

Change-Id: I6df3ea844ef1602680bb1feb8f0a63259996b1ad